+++ netomaton/evolution.py
@@ -153,8 +153,9 @@ def evolve(network, initial_conditions=None, activity_rule=None, timesteps=None,
                     a given input will be cached, and used on subsequent invocations of the rule; this can result
                     in a significant improvement to execution speed if the rule is expensive to invoke; NOTE: this
                     should only be set to True for rules which do not store any state upon invocation (default is False)
-    :param memoization_key: a user-specified key for memoization; by default, no memoization is key is configured, and a
-                            memoization key is expected if `memoize` is True; must be a `MemoizationKey` instance
+    :param memoization_key: a user-specified key for memoization; by default, an appropriate memoization key is
+                            used depending on the presence or absence of a rotation system on the network;
+                            must be a `MemoizationKey` instance; this argument is relevant only when `memoize` is `True`
                             (default is None)
     :return: a trajectory, which is a list of States, where each State defines the network and its activities for a
              particular timestep
@@ -179,11 +180,13 @@ def evolve(network, initial_conditions=None, activity_rule=None, timesteps=None,
     prev_network = network
 
     memo_table = {}
+    if memoize and memoization_key is None:
+        memoization_key = _UnorderedMemoizationKey() if network.rotation_system is None else _OrderedMemoizationKey()
 
     # create a neighbourhood map for efficiency if the network never changes (i.e. no topology rule)
     neighbourhood_map = None
     if topology_rule is None:
-        neighbourhood_map = _build_neighbourhood_map(network)
+        neighbourhood_map = _get_neighbourhood_map(network)
 
     t = 1
     while True:
@@ -198,19 +201,19 @@ def evolve(network, initial_conditions=None, activity_rule=None, timesteps=None,
         if update_order is UpdateOrder.ACTIVITIES_FIRST:
             added_nodes, removed_nodes = evolve_activities(activity_rule, t, inp, curr_activities, prev_activities,
                                                            trajectory, prev_network, past, perturbation, persist_activities,
-                                                           memoize, memo_table, memoization_key, neighbourhood_map)
+                                                           memoize, memo_table, neighbourhood_map, memoization_key)
             curr_network = evolve_topology(topology_rule, t, curr_activities, prev_network,
                                            trajectory, copy_network, persist_network, added_nodes, removed_nodes)
             if added_nodes or removed_nodes:
                 # re-build the neighbourhood map, since the network has changed
-                neighbourhood_map = _build_neighbourhood_map(curr_network)
+                neighbourhood_map = _get_neighbourhood_map(curr_network)
 
         elif update_order is UpdateOrder.TOPOLOGY_FIRST:
             curr_network = evolve_topology(topology_rule, t, prev_activities, prev_network,
                                            trajectory, copy_network, persist_network)
             # added and removed nodes are ignore in this case
             evolve_activities(activity_rule, t, inp, curr_activities, prev_activities, trajectory, curr_network, past,
-                              perturbation, persist_activities, memoize, memo_table, memoization_key, neighbourhood_map)
+                              perturbation, persist_activities, memoize, memo_table, neighbourhood_map, memoization_key)
 
         elif update_order is UpdateOrder.SYNCHRONOUS:
             # TODO create test
@@ -218,7 +221,7 @@ def evolve(network, initial_conditions=None, activity_rule=None, timesteps=None,
                                            trajectory, copy_network, persist_network)
             # added and removed nodes are ignore in this case
             evolve_activities(activity_rule, t, inp, curr_activities, prev_activities, trajectory, prev_network, past,
-                              perturbation, persist_activities, memoize, memo_table, memoization_key, neighbourhood_map)
+                              perturbation, persist_activities, memoize, memo_table, neighbourhood_map, memoization_key)
 
         else:
             raise Exception("unsupported update_order: %s" % update_order)
@@ -234,13 +237,14 @@ def evolve(network, initial_conditions=None, activity_rule=None, timesteps=None,
 
 
 def evolve_activities(activity_rule, t, inp, curr_activities, prev_activities, trajectory, network,
-                      past, perturbation, persist_activities, memoize, memoization_table, memoization_key, neighbourhood_map):
+                      past, perturbation, persist_activities, memoize, memoization_table, neighbourhood_map,
+                      memoization_key):
     added_nodes = []
     removed_nodes = []
     if activity_rule:
         added_nodes, removed_nodes = do_activity_rule(t, inp, curr_activities, prev_activities, network,
                                                       activity_rule, past, perturbation, memoize, memoization_table,
-                                                      memoization_key, neighbourhood_map)
+                                                      neighbourhood_map, memoization_key)
 
         if added_nodes:
             for state, outgoing_links, node_label in added_nodes:
@@ -253,13 +257,13 @@ def evolve_activities(activity_rule, t, inp, curr_activities, prev_activities, t
 
 
 def do_activity_rule(t, inp, curr_activities, prev_activities, network, activity_rule, past, perturbation,
-                     memoize, memoization_table, memoization_key, neigh_map):
+                     memoize, memoization_table, neigh_map, memoization_key):
     added_nodes = []
     removed_nodes = []
 
     for node_label in network.nodes:
         incoming_connections = network.in_edges(node_label)
-        neighbour_labels = neigh_map[node_label] if neigh_map is not None else [i for i in incoming_connections]
+        neighbour_labels = _get_neighbour_labels(node_label, neigh_map, network, incoming_connections)
         current_activity = prev_activities[node_label]
         neighbourhood_activities = [prev_activities[neighbour_label] for neighbour_label in neighbour_labels]
         node_in = None if inp == "__timestep__" else inp[node_label] if _is_indexable(inp) else inp
@@ -287,9 +291,18 @@ def do_activity_rule(t, inp, curr_activities, prev_activities, network, activity
     return added_nodes, removed_nodes
 
 
+def _get_neighbour_labels(node_label, neighbourhood_map, network, incoming_connections):
+    if neighbourhood_map is not None:
+        return neighbourhood_map[node_label]
+    elif network.rotation_system is not None:
+        return network.rotation_system[node_label]
+    else:
+        return [i for i in incoming_connections]
+
+
 def _get_memoized(activity_rule, ctx, memoization_table, memoization_key):
-    key = memoization_key.to_key(ctx) if memoization_key is not None else None
-    if key is not None and key in memoization_table:
+    key = memoization_key.to_key(ctx)
+    if key in memoization_table:
         return memoization_table[key]
     else:
         result = activity_rule(ctx)
@@ -301,6 +314,51 @@ def _get_memoized(activity_rule, ctx, memoization_table, memoization_key):
 class MemoizationKey:
     def to_key(self, ctx):
         pass
+
+
+class _OrderedMemoizationKey(MemoizationKey):
+    """
+    An implementation of a memoization key that simply converts the context's neighbourhood activities into
+    a tuple. This assumes that the ordering of the nodes in the neighbourhood matters, and that the ordering has
+    been specified, such as by setting a Rotation System on the network. For example, a CTRBL rule, or an ECA rule,
+    would require a Rotation System on the network.
+    """
+    def to_key(self, ctx):
+        if ctx.node_label not in ctx.neighbour_labels:
+            # include the current activity if the current node is not in the node's neighbourhood, in case the rule
+            #  depends on the current node's activity
+            # If the rule doesn't depend on the current node's activity, then we're adding cache entries
+            #  unnecessarily
+            #  e.g. [1,0,2], [1,3,2], [1,5,2], etc. all map to 9, so all we'd really need is a cache entry
+            #  with [1,2]->9, but instead we end up adding the entries [0,1,2]->9, [3,1,2]->9, [5,1,2]->9, etc.;
+            #  in this case, the user should supply their own memoization key implementation that doesn't incorporate
+            #  the current node's activity
+            return tuple([ctx.current_activity] + ctx.neighbourhood_activities)
+        return tuple(ctx.neighbourhood_activities)
+
+
+class _UnorderedMemoizationKey(MemoizationKey):
+    """
+    An implementation of a memoization key that simply converts the context's neighbourhood activities into
+    a frozenset of item counts. This assumes that the ordering of the nodes in the neighbourhood doesn't matter.
+    For example, a totalistic 1D CA rule does not require an ordering of the neighbourhood to be specified.
+    It would be less efficient to just return a tuple of the activities here, since we'd end up with needless
+    entries in the cache. The rule doesn't depend on node order, so states like (1,0,1) and (0,1,1) should result
+    in the same cache hit, and evaluate to the same result.
+    """
+    def to_key(self, ctx):
+        neigh_key = frozenset(collections.Counter(ctx.neighbourhood_activities).items())
+        if ctx.node_label not in ctx.neighbour_labels:
+            # include the current activity if the current node is not in the node's neighbourhood, in case the rule
+            #  depends on the current node's activity
+            # If the rule doesn't depend on the current node's activity, then we're adding cache entries
+            #  unnecessarily
+            #  e.g. [1,0,2], [2,3,1], [2,5,1], etc. all map to 9, so all we'd really need is a cache entry
+            #  with {1,2}->9, but instead we end up adding the entries (0,{1,2})->9, (3,{1,2})->9, (5,{1,2})->9, etc.;
+            #  in this case, the user should supply their own memoization key implementation that doesn't incorporate
+            #  the current node's activity
+            return ctx.current_activity, neigh_key
+        return neigh_key
 
 
 def evolve_topology(topology_rule, t, activities, prev_network, trajectory, copy_network,
@@ -347,7 +405,9 @@ def _get_past_activities(past_conditions, trajectory, t):
     return None
 
 
-def _build_neighbourhood_map(network):
+def _get_neighbourhood_map(network):
+    if network.rotation_system is not None:
+        return network.rotation_system
     return {node_label: [i for i in network.in_edges(node_label)] for node_label in network.nodes}