+++ tests/test_game_of_life.py
@@ -5,7 +5,7 @@ from .rule_test import *
 class TestGameOfLifeRule(RuleTest):
 
     def test_gol(self):
-        network = ntm.topology.cellular_automaton2d(rows=60, cols=60, r=1, neighbourhood='Moore')
+        network = ntm.topology.cellular_automaton2d(rows=60, cols=60, r=1, neighbourhood="Moore")
         expected = self._convert_to_list_of_lists("game_of_life.ca")
 
         initial_conditions = ntm.init_simple2d(60, 60)
@@ -40,7 +40,7 @@ class TestGameOfLifeRule(RuleTest):
         np.testing.assert_equal(expected, activities)
 
     def test_gol_memoized(self):
-        network = ntm.topology.cellular_automaton2d(rows=60, cols=60, r=1, neighbourhood='Moore')
+        network = ntm.topology.cellular_automaton2d(rows=60, cols=60, r=1, neighbourhood="Moore")
         expected = self._convert_to_list_of_lists("game_of_life.ca")
 
         initial_conditions = ntm.init_simple2d(60, 60)
@@ -69,8 +69,7 @@ class TestGameOfLifeRule(RuleTest):
         initial_conditions[2417] = 1
 
         trajectory = ntm.evolve(initial_conditions=initial_conditions, network=network, timesteps=180,
-                                activity_rule=ntm.rules.game_of_life_rule, memoize=True,
-                                memoization_key=ntm.rules.CenteringMemoizationKey())
+                                activity_rule=ntm.rules.game_of_life_rule, memoize=True)
 
         activities = ntm.get_activities_over_time_as_list(trajectory)
         np.testing.assert_equal(expected, activities)