+++ tests/test_network.py
@@ -1,6 +1,7 @@
 import netomaton as ntm
 from .rule_test import *
 import networkx as nx
+import pytest
 
 
 class TestNetwork(RuleTest):
@@ -399,3 +400,53 @@ class TestNetwork(RuleTest):
             [0.0, 0.0, 2.0],
             [0.0, 2.0, 0.0],
         ], M)
+
+    def test_rotation_system(self):
+        network = ntm.Network()
+        network.add_edge("A", "B")
+        network.add_edge("A", "A")
+        network.add_edge("B", "A")
+        network.add_edge("B", "B")
+        network.add_edge("C", "B")
+        network.add_edge("B", "C")
+        network.add_edge("C", "C")
+
+        try:
+            network.rotation_system = {
+                "A": ("B", "A"),
+                "B": ("A", "B", "C"),
+                "C": ("B", "C")
+            }
+        except Exception as e:
+            self.fail("raised Exception unexpectedly: %s" % e.args)
+
+    def test_rotation_system_no_self_connections(self):
+        network = ntm.Network()
+        network.add_edge("A", "B")
+        network.add_edge("B", "A")
+        network.add_edge("C", "B")
+        network.add_edge("B", "C")
+
+        try:
+            network.rotation_system = {
+                "A": ("B",),
+                "B": ("A", "C"),
+                "C": ("C",)
+            }
+        except Exception as e:
+            self.fail("raised Exception unexpectedly: %s" % e.args)
+
+    def test_rotation_system_illegal(self):
+        network = ntm.Network()
+        network.add_edge("A", "B")
+        network.add_edge("B", "A")
+        network.add_edge("C", "B")
+        network.add_edge("B", "C")
+
+        with pytest.raises(Exception) as e:
+            network.rotation_system = {
+                "A": ("B", "A"),
+                "B": ("A", "B", "C"),
+                "C": ("B", "C", "D")
+            }
+        self.assertEqual(e.value.args, ("the incoming node 'D' is not in the network",))