+++ tests/test_wireworld_rule.py
@@ -13,7 +13,7 @@ class TestWireworldRule(RuleTest):
 
     def test_wireworld_diodes_memoized(self):
         expected = self._convert_to_matrix2d("wireworld_diodes.ca")
-        actual = self._evolve_wireworld(expected, memoize=True, memoization_key=rules.CenteringMemoizationKey())
+        actual = self._evolve_wireworld(expected, memoize=True)
         np.testing.assert_equal(expected, actual)
 
     def test_wireworld_xor(self):
@@ -23,15 +23,15 @@ class TestWireworldRule(RuleTest):
 
     def test_wireworld_xor_memoized(self):
         expected = self._convert_to_matrix2d("wireworld_xor.ca")
-        actual = self._evolve_wireworld(expected, memoize=True, memoization_key=rules.CenteringMemoizationKey())
+        actual = self._evolve_wireworld(expected, memoize=True)
         np.testing.assert_equal(expected, actual)
 
     @staticmethod
-    def _evolve_wireworld(expected, memoize=False, memoization_key=None):
+    def _evolve_wireworld(expected, memoize=False):
         steps, rows, size = expected.shape
         initial_conditions = np.array(expected[0]).reshape(rows * size).flatten()
         network = cellular_automaton2d(rows=rows, cols=size, neighbourhood="Moore")
         trajectory = evolve(initial_conditions=initial_conditions, network=network, timesteps=steps,
-                            activity_rule=rules.wireworld_rule, memoize=memoize, memoization_key=memoization_key)
+                            activity_rule=rules.wireworld_rule, memoize=memoize)
         activities = get_activities_over_time_as_list(trajectory)
         return np.array(activities).reshape((steps, rows, size))